// Client/front-end/src/Models/Producatordto.ts
export class ProducatorDTO {
     producatorId: number;
     producatorIdCt: number;
     dataInceput: string;
     dataSfarsit: string;
     nume: string;
     adresa: string;
     mmId: number;
    iData: string
    iCalculator: string
    iUtilizator: string
    uData: string
    uCalculator: string
    uUtilizator: string
    rv: string
}